// frontend/src/api/models/handlerCustomerListResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerCustomerListResponseLevel } from './handlerCustomerListResponseLevel'
import type { HandlerCustomerListResponseStatus } from './handlerCustomerListResponseStatus'
import type { HandlerCustomerListResponseType } from './handlerCustomerListResponseType'

/**
 * Customer list item with basic information
 */
export interface HandlerCustomerListResponse {
  city?: string
  code?: string
  created_at?: string
  email?: string
  id?: string
  level?: HandlerCustomerListResponseLevel
  name?: string
  phone?: string
  province?: string
  short_name?: string
  sort_order?: number
  status?: HandlerCustomerListResponseStatus
  type?: HandlerCustomerListResponseType
}