// frontend/src/api/models/handlerCustomerListResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCustomerListResponseStatus =
  (typeof HandlerCustomerListResponseStatus)[keyof typeof HandlerCustomerListResponseStatus]

export const HandlerCustomerListResponseStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const