// frontend/src/api/models/handlerCustomerListResponseLevel.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCustomerListResponseLevel =
  (typeof HandlerCustomerListResponseLevel)[keyof typeof HandlerCustomerListResponseLevel]

export const HandlerCustomerListResponseLevel = {
  normal: 'normal',
  silver: 'silver',
  gold: 'gold',
  platinum: 'platinum',
  vip: 'vip',
} as const