// frontend/src/api/models/handlerCustomerListResponseType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCustomerListResponseType =
  (typeof HandlerCustomerListResponseType)[keyof typeof HandlerCustomerListResponseType]

export const HandlerCustomerListResponseType = {
  individual: 'individual',
  organization: 'organization',
} as const