// frontend/src/api/models/getPartnerCustomersParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { GetPartnerCustomersLevel } from './getPartnerCustomersLevel'
import type { GetPartnerCustomersOrderDir } from './getPartnerCustomersOrderDir'
import type { GetPartnerCustomersStatus } from './getPartnerCustomersStatus'
import type { GetPartnerCustomersType } from './getPartnerCustomersType'

export type GetPartnerCustomersParams = {
  /**
   * Search term (name, code, phone, email)
   */
  search?: string
  /**
   * Customer status
   */
  status?: GetPartnerCustomersStatus
  /**
   * Customer type
   */
  type?: GetPartnerCustomersType
  /**
   * Customer level
   */
  level?: GetPartnerCustomersLevel
  /**
   * City
   */
  city?: string
  /**
   * Province
   */
  province?: string
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Order by field
   */
  order_by?: string
  /**
   * Order direction
   */
  order_dir?: GetPartnerCustomersOrderDir
}