// frontend/src/api/models/getPartnerCustomersStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetPartnerCustomersStatus =
  (typeof GetPartnerCustomersStatus)[keyof typeof GetPartnerCustomersStatus]

export const GetPartnerCustomersStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const