// frontend/src/api/models/getPartnerCustomersLevel.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetPartnerCustomersLevel =
  (typeof GetPartnerCustomersLevel)[keyof typeof GetPartnerCustomersLevel]

export const GetPartnerCustomersLevel = {
  normal: 'normal',
  silver: 'silver',
  gold: 'gold',
  platinum: 'platinum',
  vip: 'vip',
} as const