// frontend/src/api/models/getPartnerCustomersType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetPartnerCustomersType =
  (typeof GetPartnerCustomersType)[keyof typeof GetPartnerCustomersType]

export const GetPartnerCustomersType = {
  individual: 'individual',
  organization: 'organization',
} as const